full name in view whole week, also fixed viewwholeweek loading forever, also when logout navigating to login...

+++ src/app/(dashboard)/roles/page.js
@@ -44,6 +44,7 @@ const Roles = () => {
 
   // Function to delete an existing role
   const handleDeleteRole = async (roleId) => {
+    if (!window.confirm(t("deleteConfirmation"))) return;
     setLoading(true);
     deleteRole(roleId)
       .then((res) => {
